model_02_ReportGuideNet/PlainNet/requirements.txt
# Install torch & torchvision with the correct CUDA wheel separately.
# Example (CUDA 12.1):
# pip install torch==2.2.2+cu121 torchvision==0.17.2+cu121 --extra-index-url https://download.pytorch.org/whl/cu121

numpy<2
pyyaml
pillow
scikit-learn
tqdm
